TM-4: combined TaskList and create-task components into one model, and moved the folders into task-feature folder
