+++ FirstApp/src/app/tasks-feature/services/task.spec.ts
@@ -3,7 +3,7 @@ import { TestBed } from '@angular/core/testing';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 
 import { TaskService } from './task';
-import { Task } from '../models/task';
+import { Task } from '../../models/task';
 
 describe('TaskService', () => {
   let service: TaskService;
